Moving top level reducers/sagas into a data dir
