+++ src/data/sagas.js
@@ -1,5 +1,5 @@
 import { all } from 'redux-saga/effects';
-import { saga as accountSettingsSaga } from './account-settings';
+import { saga as accountSettingsSaga } from '../account-settings';
 
 export default function* rootSaga() {
   yield all([accountSettingsSaga()]);
